add ability to display login error
This change allows the view controller to access the login object that
the q creates. The indices that are reported to the delegate have been
reduced by 1, so that the delegate can use the index directly on the
object store. Note that we must check for a -1 (which is the login
object failed) else we might ask an array for an object at index -1.
DANGER.